refactor(VZOG-11): 강의 상세 컴포넌트 분리
